refactor: create type and const
