--- src/components/HelloWorld.vue
@@ -1,45 +1,0 @@
-<script setup lang="ts">
-import { ref } from 'vue'
-
-const count = ref(0)
-
-defineProps<{ msg: string }>()
- function hundleclick() {
-   count.value = count.value + 2
-   console.log(count)
- }
-
-
-</script>
-
-<template>
-  <h1>{{ msg }}</h1>
-
-  <div class="card">
-    <button type="button" @click="hundleclick">count is {{ count }}</button>
-    <p>
-      Edit
-      <code>components/HelloWorld.vue</code> to test HMR
-    </p>
-  </div>
-
-  <p>
-    Check out
-    <a href="https://vuejs.org/guide/quick-start.html#local" target="_blank"
-      >create-vue</a
-    >, the official Vue + Vite starter
-  </p>
-  <p>
-    Install
-    <a href="https://github.com/vuejs/language-tools" target="_blank">Volar</a>
-    in your IDE for a better DX
-  </p>
-  <p class="read-the-docs">Click on the Vite and Vue logos to learn more</p>
-</template>
-
-<style scoped>
-.read-the-docs {
-  background-color: #888;
-  color: #888;
-}
-</style>
