+++ src/consts/index.ts
@@ -1,0 +1,11 @@
+import { WEEK_DAY_TYPE } from "../types";
+
+export const WEEK_DAY: Record<number, WEEK_DAY_TYPE> = {
+    0: 'Sun',
+    1: 'Mon',
+    2: 'Tue',
+    3: 'Wed',
+    4: 'Thu',
+    5: 'Fri',
+    6: 'Sat',
+};
