+++ src/types/index.ts
@@ -1,0 +1,1 @@
+export type WEEK_DAY_TYPE = 'Sun' | 'Mon' | 'Tue' | 'Wed' | 'Thu' | 'Fri' | 'Sat';
